+++ openm/omc/DimensionSymbol.cpp
@@ -37,7 +37,7 @@ void DimensionSymbol::to_column_name(const string & i_ownerName, const list<Dime
 {
     assert(io_dim);
     size_t maxlen = std::min<size_t>(short_name_max_length, OM_NAME_DB_MAX);
-    string colName = openm::toAlphaNumeric(io_dim->short_name, maxlen);  // make dimension name alphanumeric and truncate it to 255 chars
+    string colName = openm::toAlphaNumeric(io_dim->short_name, (int)maxlen);  // make dimension name alphanumeric and truncate it to 255 chars
 
     for (auto pIt = i_dimLst.cbegin(); pIt != i_dimLst.cend() && *pIt != io_dim; ++pIt) {
         if (colName == (*pIt)->short_name) {
@@ -228,7 +228,7 @@ string DimensionSymbol::heuristic_short_name(void) const
     }
 
     // Make name alphanumeric and truncate it to maximum length.
-    hn = openm::toAlphaNumeric(hn, short_name_max_length);
+    hn = openm::toAlphaNumeric(hn, (int)short_name_max_length);
 
     // trim off trailing "_" if present
     //if (hn.ends_with("_")) { // c++20
