+++ openm/omc/EnumerationSymbol.cpp
@@ -52,7 +52,7 @@ void EnumerationSymbol::populate_metadata(openm::MetaModelHolder & metaRows)
         typeDic.typeId = type_id;
         typeDic.name = name;
         typeDic.dicId = dicId;          // see enum kind_of_type for possible values
-        typeDic.totalEnumId = pp_size();
+        typeDic.totalEnumId = (int)pp_size();
         metaRows.typeDic.push_back(typeDic);
     }
